function/requirements.txt
lxml
cssselect
google-cloud-firestore
requests